GitLab sync: Include inherited members when checking for write permissions (#1537)
We're currently blocked from using the plugin because after connecting
to our private GitLab server the plugin refuses to let us write any
changes (even though it was able to create a commit with the new json
file).

After some debugging we found that the GitLab API was returning a 404
error when trying to fetch the member data for the project. This is
because our GitLab repository inherits the members from another group
and it doesn't have any direct members itself. To get all the members
including the inherited ones you need to fetch another endpoint.

By including the `includeInherited` option in the calls to
`GroupMembers` and `ProjectMembers` from the `@gitbeaker/browser`
package this issue will be resolved for us as it will hit the correct
endpoint. According to GitLab's own documentation this should include
all members at all times
https://docs.gitlab.com/ee/api/members.html#list-all-members-of-a-group-or-project-including-inherited-and-invited-members,
so it should not have any impact on existing use cases.

Co-authored-by: Robin Ambachtsheer <robin.ambachtsheer@mollie.com>