fix(genericVersioned): save last synced state when tokens update (#1839)
Updates the `lastSyncedState` when generic versioned tokens update, and sets the `tokenSetOrder` property in the metadata based on the order of keys in the JSON.

The genericVersioned provider had a bug where the `metadata` property was not set correctly in the `state` leading to false-positives when checking for state changes. This solution taken from #1108 and #1395.