Supernova.io Integration (#1809)
This PR adds support for the Supernova data provider. The provider works
analogous to every other git-type provider on the platform, minus the
support for branches which is not a concept Supernova has.

An upcoming docs page for this integration [can be found
here](https://learn.supernova.io/latest/design-systems/add-design-system-content/tokens-studio.html).

This integration is currently not testable with the production
environments of Supernova (we need to merge prod API support for it
first), but should be available in next few days as well. In a meantime,
I'll share access for the development realm privately.

Start by adding a new provider:

<img width="782" alt="tokens"
src="https://user-images.githubusercontent.com/7470393/232052291-4ec4d1f3-0fe1-4fa6-b3f6-2d53dd5a1bf1.png">

---------

Co-authored-by: Jiri Trecak <jiri@supernova.io>
Co-authored-by: Jan Six <six7@github.com>
Co-authored-by: Jan Six <six.jan@gmail.com>