Perf fixes (#2092)
Removed migrateTokens call. This was something we added to migrate
border tokens to borderColor tokens. I's likely ok to remove this, it's
been around 6 months since we introduced the border tokens or more. This
was responsible for a lot of calls to Updating plugin data... which
caused everything to be rather slow.

I added a few checks to document changes. This is basically responsible
for checking if other users are changing something to applied tokens,
and then we would invalidate and re-fetch those nodes. However, it also
ran on all document changes, including internal ones and from the user.
I removed this, and it should lead to less calls and faster execution
time.

Also fixes a major issue that caused us to iterate over all variable
references _on every node_ resulting in poor performance.