fix apply performance (#2172)
Fixes https://github.com/tokens-studio/figma-plugin/issues/2129

Removes `updatePluginData` calls from the `update` call, as we only used
that for migration purposes. We would still do that when users apply
tokens or remap tokens, but we won't do anymore on every Apply to.. call

* `src/plugin/asyncMessageHandlers/update.ts`: The `updatePluginData`
function is no longer used in the `update.ts` file, so the import
statement and call to the function have been removed. (remove
`updatePluginData` references in `update.ts`)




https://github.com/tokens-studio/figma-plugin/assets/4548309/57759166-42be-4d1e-878a-89ecc63301f7