Change font-size input field position in Typography composite token modal Current: Font size is after line-height
Problem: Line-height is often dependent on font-size, so it would help to first set the font-size and then the line-height.

Solution: Change the font-size input field position in the Typography composite token modal.

PS: this is my first ever contribution to an open-source project, please be gentle when looking at naming conventions for branches/commits