W3C DTCG format (#2449)
<!--
  Notes for authors:
  - Provide context with minimal words, keep it concise
  - Mark as a draft for work in progress PRs
  - Once ready for review, notify others in #code-reviews
- Remember, the review process is a learning opportunity for both
reviewers and authors, it's a way for us to share knowledge and avoid
silos.
-->

### Why does this PR exist?

Closes #999

Introduces support for the DTCG format as in supporting `$value`,
`$type` and `$description` in both input and output.

### What does this pull request do?

Introduces a setting that, when changed, writes and reads tokens to the
`$-prefixed` props instead of the legacy ones we supported before. We'll
still support both, but this will allow the new format to be chosen.

This is part of a larger epic where in the end we'll allow users to
choose the format (possibly on the provider). Another thought is that we
could just automatically detect the format and then use that, but then
offer users a one-time "migrate to new" button where we'd convert their
tokens.

### Testing this change

<!--
Describe how this change can be tested. Are there steps required to get
there? Explain what's required so a reviewer can test these changes
locally.

  If you have a review link available, add it here.
-->

### Additional Notes (if any)

<!--
  Add any other context or screenshots about the pull request
-->