refactor applying tokens + add support for binding variables to effects (#2498)
### Why does this PR exist?

Closes https://github.com/tokens-studio/figma-plugin/issues/2443

Figma recently added support for binding variables to effects such as
shadows. This PR adds support for this.

### What does this pull request do?

Changes seem big - which is because this PR includes a refactor of our
update function to make it WAY easier to parse and add new features. The
big `setValuesOnNode` file has been split into many.

I also added a way to retrieve tokens and references from the plugin
side from a central store instead of passing it through for every
function, i could imagine this was another sort for memory leaks.

This was required as otherwise we would not have a proper way to detect
the connected variables for a value in a composite token, for example a
color in a shadow, or any dimension value in any of the x, y, spread,
values of a shadow. Once Figma release typography variables this will
come in very handy as we'd already be prepared.

### Testing this change

Apply a couple of tokens. Composition tokens, tokens using references.
Create styles, create variables, give it a gooooood test.