Fix/2596 text overflow in set list (#2602)
<!--
  Notes for authors:
  - Provide context with minimal words, keep it concise
  - Mark as a draft for work in progress PRs
  - Once ready for review, notify others in #code-reviews
- Remember, the review process is a learning opportunity for both
reviewers and authors, it's a way for us to share knowledge and avoid
silos.
-->

### Why does this PR exist?

Closes #2596  <!-- link the related issue -->

Before: 


![image](https://github.com/tokens-studio/figma-plugin/assets/23120116/51ab1437-141c-4051-8bcd-2fdd2646a2a0)


After: 

![image](https://github.com/tokens-studio/figma-plugin/assets/23120116/42d3b8f6-2294-4eb3-bd31-a6d8e29ae2c9)


<!--
Describe the problem you're addressing and the rationale behind this PR.
-->

### What does this pull request do?

<!--
Detailed summary of the changes, including any visual or interactive
updates.
For UI changes, add before/after screenshots. For interactive elements,
consider including a video or an animated gif.
Explain some of the choices you've made in the PR, if they're not
obvious.
-->

### Testing this change

<!--
Describe how this change can be tested. Are there steps required to get
there? Explain what's required so a reviewer can test these changes
locally.

  If you have a review link available, add it here.
-->

### Additional Notes (if any)

<!--
  Add any other context or screenshots about the pull request
-->