2576 20 styles variables options should persist between sessions (#2612)
<!--
  Notes for authors:
  - Provide context with minimal words, keep it concise
  - Mark as a draft for work in progress PRs
  - Once ready for review, notify others in #code-reviews
- Remember, the review process is a learning opportunity for both
reviewers and authors, it's a way for us to share knowledge and avoid
silos.
-->

### Why does this PR exist?
The current implementation does not store styles and variables export
options and the options expires when plugin launches.

Closes #2576 <!-- link the related issue -->

<!--
Describe the problem you're addressing and the rationale behind this PR.
-->

### What does this pull request do?
Add effect functions to setting reducers to trigger `setUI` function to
store export options.

<!--
Detailed summary of the changes, including any visual or interactive
updates.
For UI changes, add before/after screenshots. For interactive elements,
consider including a video or an animated gif.
Explain some of the choices you've made in the PR, if they're not
obvious.
-->

### Testing this change

<!--
Describe how this change can be tested. Are there steps required to get
there? Explain what's required so a reviewer can test these changes
locally.

  If you have a review link available, add it here.
-->

### Additional Notes (if any)

<!--
  Ad


https://github.com/tokens-studio/figma-plugin/assets/103296157/97d45670-89e4-4d0b-ac87-b0d38c6e5a41

d any other context or screenshots about the pull request
-->

---------

Co-authored-by: Jan Six <six.jan@gmail.com>