#2620 [2.0] inspect panel actions dont have spacing (#2626)
<!--
  Notes for authors:
  - Provide context with minimal words, keep it concise
  - Mark as a draft for work in progress PRs
  - Once ready for review, notify others in #code-reviews
- Remember, the review process is a learning opportunity for both
reviewers and authors, it's a way for us to share knowledge and avoid
silos.
-->

### Why does this PR exist?

Closes #2620 <!-- link the related issue -->

<!--
Describe the problem you're addressing and the rationale behind this PR.
-->

### What does this pull request do?

Added gap to the buttons - 8px following HiG and Apple. 16px in the
original ticket is too much.
Changed flex behaviour so buttons are always visible
Set the 'remove selected' button to danger variant

After: 

![Kapture 2024-04-11 at 12 36
44](https://github.com/tokens-studio/figma-plugin/assets/23120116/3241b542-a275-4e9c-bc77-e42e376b87d7)


<!--
Detailed summary of the changes, including any visual or interactive
updates.
For UI changes, add before/after screenshots. For interactive elements,
consider including a video or an animated gif.
Explain some of the choices you've made in the PR, if they're not
obvious.
-->

### Testing this change

<!--
Describe how this change can be tested. Are there steps required to get
there? Explain what's required so a reviewer can test these changes
locally.

  If you have a review link available, add it here.
-->

### Additional Notes (if any)

<!--
  Add any other context or screenshots about the pull request
-->