fix re-render of manage styles modal (sets) + polish (#2638)
### Why does this PR exist?

The export tab for variables / styles had a re-render that was caused by
the dialog coming from a hook. this fixes this.

### What does this pull request do?

In addition to the re-render, optimizes styling of the `Options` dialog
to not use as much vertical space.

The changes in this PR seem large, but it's really just moving the
render part of the hook out to its own files, and then add some styling.

The other change this introduces is to treat "color" as a toggle. We
dont want users to create both styles and variables for colors at the
same time, this is always unintended. I now changed it so it is a
toggle, if color is set true, the other will be set to false.

### Testing this change

Interact with the manage styles and variables modal

### Additional Notes (if any)

Before:


https://github.com/tokens-studio/figma-plugin/assets/4548309/bbc48686-9677-45ab-9a76-aa630b3c2691


After:


https://github.com/tokens-studio/figma-plugin/assets/4548309/4735501d-e59e-4738-9312-79a3e1689157