2.0 Fix issue with token key order (#2659)
### Why does this PR exist?

Closes https://github.com/tokens-studio/figma-plugin/issues/2653

In the prior version of the 2.0 beta we would change the token key order
unintentionally as we are now keeping track of the w3c defined key
naming. However, we dont want to rearrange the keys of our users.

### What does this pull request do?

Instead of spreading the new values on the internal tokens, we will just
re-assign the values.

### Testing this change

#### Before


https://github.com/tokens-studio/figma-plugin/assets/4548309/0608a4c0-653e-4236-abfc-68286fe8d93a

#### After


https://github.com/tokens-studio/figma-plugin/assets/4548309/d514a707-8655-4ef0-a8e3-267498dd95e0