2.0: refactor changing aliases in state and fix loading bar bug  (#2704)
### Why does this PR exist?

Closes https://github.com/tokens-studio/figma-plugin/issues/2642

There was a loading bar that appeared when you renamed a token group,
that never disappeared.

### What does this pull request do?

The change fixing this issue is actually fairly small, I just removed
the job that we had for this. The job itself was not an async operation,
so there was no need for the loading bar. I assume it could have been a
race condition that caused the start job and the stop job to never be
submitted.

However, doing this I refactored how we call that `updateAlias` function
as I discovered we had quite a mess there including a lot of logic
directly in state which was not tested, a `updateOtherAlias` function
which had no use. It basically was required as previously we didnt
update modifier props when we update the aliases, but now I changed that
to just happen before we perform the alias change, and added tests for
that function, plus moved it to a dedicated new `utils` directory in the
store where we should move all logic that currently clutters the
tokenState.

### Testing this change

Test coverage is pretty good with that one, but here's a video
demonstrating the change:


https://github.com/tokens-studio/figma-plugin/assets/4548309/ea4a7b7b-8f7b-4a39-9de0-f4d4e89f7ee6

### Additional Notes (if any)

We should really start to clean up our `tokenState` and move all the
logic into dedicated files that we can test in isolation.