2.0: fix startup (#2725)
### Why does this PR exist?

Fixes https://github.com/tokens-studio/figma-plugin/issues/2719

### What does this pull request do?

Previously `addLicenseKey` was a side-effect in state. This caused us to
not await the response of this, something in the upgrade to 2.0 must've
caused this to become apparent.

I now changed it so that addLicenseKey is a dedicated function that we
properly await. Doing so the plugin now awaits license key response
before we move on to other parts of the startup, also fixing our launch
darkly issues.

### Before

https://github.com/tokens-studio/figma-plugin/assets/4548309/b7722b8a-2e2b-47ba-9b6f-b5b6c955eadd

### After

https://github.com/tokens-studio/figma-plugin/assets/4548309/cec4d03f-051e-40f4-a149-ea26b731bbc6